% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAnnotation.R
\name{addHelpDoc}
\alias{addHelpDoc}
\title{Add README.Rmd with Help}
\usage{
addHelpDoc(out = "Supporting_Macros/README.Rmd")
}
\arguments{
\item{out}{path to output help file to}
}
\description{
Add README.Rmd with Help
}

