% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drat.R
\name{addToDrat}
\alias{addToDrat}
\title{Add package to drat repo}
\usage{
addToDrat(pkg, cleanup = TRUE, ...)
}
\arguments{
\item{pkg}{name of the package}

\item{cleanup}{boolean indicating if built files should be cleaned up}

\item{...}{additional parameters to be passed to \code{\link{insertPackage}}
for binaries}
}
\description{
Add package to drat repo
}

