% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{AlteryxGraph2}
\alias{AlteryxGraph2}
\title{Alteryx Graph Function}
\usage{
AlteryxGraph2(expr, nOutput = 1, width = 576, height = 576, ...)
}
\arguments{
\item{expr}{expression to generate graph}

\item{nOutput}{output connection number}

\item{width}{width}

\item{height}{height}

\item{...}{additional arguments}
}
\description{
Alteryx Graph Function
}

