% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{copyAlteryxRPlugin}
\alias{copyAlteryxRPlugin}
\title{Copy predictive macros, samples and plugins from SVN}
\usage{
copyAlteryxRPlugin(to = NULL)
}
\arguments{
\item{to}{directory to copy files to}
}
\description{
Copy predictive macros, samples and plugins from SVN
}

