% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHtmlPluginNew.R
\name{copyHtmlPlugin}
\alias{copyHtmlPlugin}
\title{Update Html Plugin}
\usage{
copyHtmlPlugin(pluginDir = ".", ayxDir = getAyxDirs())
}
\arguments{
\item{pluginDir}{plugin directory to copy from}

\item{ayxDir}{alteryx directories to copy to}
}
\description{
Update Html Plugin
}

