% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createPluginFromMacro}
\alias{createPluginFromMacro}
\title{Create Plugin from Macro}
\usage{
createPluginFromMacro(pluginDir = ".", overrides = NULL, layout = NULL,
  render_readme = FALSE, ...)
}
\arguments{
\item{pluginDir}{plugin directory}

\item{overrides}{should an override file be used}

\item{layout}{should a layout file be used}

\item{render_readme}{whether or not to render READMEs}

\item{...}{arguments to pass on to makeIcon}
}
\description{
Create Plugin from Macro
}

