% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createYXI}
\alias{createYXI}
\title{Create YXI file}
\usage{
createYXI(pluginDir = ".", toDir = ".")
}
\arguments{
\item{pluginDir}{plugin directory}

\item{toDir}{to directory}
}
\description{
Create YXI file
}

