% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installAlteryx.R
\name{downloadInstallers}
\alias{downloadInstallers}
\title{Download Installers from the Build Repo}
\usage{
downloadInstallers(buildRepo = "\\\\\\\\DEN-IT-FILE-07\\\\BuildRepo",
  to = "C:/Users/ramnath/Downloads", buildDir = NULL,
  branch = "Predictive_Dev", type = "Server", rInstaller = "RInstaller")
}
\arguments{
\item{buildRepo}{path to build repo.}

\item{to}{directory to download installers to.}

\item{buildDir}{build directory.}

\item{branch}{string indicating branch. defaults to Predictive_Dev.}

\item{type}{string indicating installer type. one of 'Server', 'Gallery',
'NonAdmin' or ''.}
}
\description{
Download Installers from the Build Repo
}

