% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAnnotation.R
\name{extractConfig}
\alias{extractConfig}
\title{Extract configuration annotation from macro}
\usage{
extractConfig(pluginName)
}
\arguments{
\item{pluginName}{path to yxmc file}
}
\description{
Extract configuration annotation from macro
}

