% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAnnotation.R
\name{extractInputOutput}
\alias{extractInputOutput}
\title{Extract annotation from Inputs}
\usage{
extractInputOutput(pluginName, type = "input")
}
\arguments{
\item{pluginName}{path to yxmc file}
}
\description{
Extract annotation from Inputs
}

