% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput.R
\name{extractQuestionConstants}
\alias{extractQuestionConstants}
\title{Extract question constants and return code that can be inserted}
\usage{
extractQuestionConstants(macro, input = "config")
}
\arguments{
\item{input}{variable to assign the question constants to}

\item{path}{to macro from which to extract question constants}
}
\description{
Extract question constants and return code that can be inserted
}

