% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/workflow.R
\name{extractRcode}
\alias{extractRcode}
\title{Save R code extracted from an Alteryx Macro}
\usage{
extractRcode(template, outFile = NULL, extractInput = NULL)
}
\arguments{
\item{template}{yxmc file to extract code from}

\item{outFile}{output file to save the updated macro to}

\item{extractInput}{question inputs automatically get extracted to list
object with this name}
}
\description{
Save R code extracted from an Alteryx Macro
}

