% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{findInputs}
\alias{findInputs}
\title{Find inputs that start with a root}
\usage{
findInputs(input, root)
}
\arguments{
\item{input}{input}

\item{root}{root}
}
\description{
Find inputs that start with a root
}

