% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{findTrueInput}
\alias{findTrueInput}
\title{Find inputs that start with a root and have the value TRUE}
\usage{
findTrueInput(input, root)
}
\description{
Find inputs that start with a root and have the value TRUE
}

