% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateConfigurationTable}
\alias{generateConfigurationTable}
\title{Generate table of configuration items}
\usage{
generateConfigurationTable(yxmcFile)
}
\arguments{
\item{yxmcFile}{path to macro}
}
\description{
Generate table of configuration items
}

