% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{getAyxSvnDirs}
\alias{getAyxSvnDirs}
\title{Get Alteryx SVN Directories}
\usage{
getAyxSvnDirs(svnDir = getOption("alteryx.svndir"))
}
\arguments{
\item{svnDir}{svn directory}
}
\description{
Get Alteryx SVN Directories
}

