% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{getAyxSvnRDirs}
\alias{getAyxSvnRDirs}
\title{Get Alteryx R SVN Directories}
\usage{
getAyxSvnRDirs(svnDir = getOption("alteryx.svndir"))
}
\arguments{
\item{svnDir}{svn directory}
}
\description{
Get Alteryx R SVN Directories
}

