% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginconfig.R
\name{getIO}
\alias{getIO}
\title{Get input/output configuration information from a macro}
\usage{
getIO(template)
}
\description{
Get input/output configuration information from a macro
}
\examples{
 template <- "~/Desktop/SNIPPETS/dev/Optimization/Supporting_Macros/Optimization.yxmc"
 x = getIO(template)
}

