% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inAlteryx}
\alias{inAlteryx}
\title{Check if the execution environment is Alteryx.}
\usage{
inAlteryx()
}
\description{
Determine if code is being executed inside an Alteryx R Tool. It utilizes the
presence of certain global variables like \code{AlteryxDataOutput} to detect
if the execution environment is Alteryx or not.
}

