% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncRcode.R
\name{insertRcode2}
\alias{insertRcode2}
\title{Syntactic sugar for insertRcode to cover the most common usage pattern.}
\usage{
insertRcode2(pluginDir = ".")
}
\description{
Syntactic sugar for insertRcode to cover the most common usage pattern.
}

