% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{installAllPackages2}
\alias{installAllPackages2}
\title{Install all packages}
\usage{

  installAllPackages2(cranRepo = "https://mran.revolutionanalytics.com/snapshot/2016-01-01/",
  ayxRepo = "https://alteryx.github.io/drat",
  buildRepo = "\\\\\\\\DEN-IT-FILE-07\\\\BuildRepo",
  branch = "Predictive_Development")
}
\description{
Install all packages
}

