% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{installAlteryx}
\alias{installAlteryx}
\title{Install Alteryx}
\usage{
installAlteryx(installers, from = "C:/Users/ramnath/Downloads")
}
\arguments{
\item{installers}{list of named installers.}

\item{from}{directory to install from.}
}
\description{
Install Alteryx
}

