% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{isOlder}
\alias{isOlder}
\title{Check if a file is older than the other}
\usage{
isOlder(x, y)
}
\arguments{
\item{x}{path to the first file}

\item{y}{path to the second file}
}
\description{
Check if a file is older than the other
}

