% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listPredictiveMacros}
\alias{listPredictiveMacros}
\title{List predictive macros.}
\usage{
listPredictiveMacros(...)
}
\arguments{
\item{...}{additional arguments to pass on to dir}
}
\description{
This returns a list of predictive macros in Alteryx. The list can be filtered
by passing additional arguments to the \code{\link{dir}} function.
}

