% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginconfig.R
\name{makePluginConfig}
\alias{makePluginConfig}
\title{Make plugin configuration xml file}
\usage{
makePluginConfig(inputs, outputs, pluginName, properties = NULL,
  helpLink = "")
}
\arguments{
\item{inputs}{inputs}

\item{outputs}{outputs}

\item{pluginName}{pluginName}

\item{properties}{properties}

\item{helpLink}{help link}
}
\description{
Make plugin configuration xml file
}
\examples{
x = list(
  inputs = list(
    list(Name = 'Input A', Abbrev = "A"),
    list(Name = 'Input B', Abbrev = "B")
  ),
  outputs = list(
    list(Name = 'Output C', Abbrev = "C")
  ),
  pluginName = 'Foo'
)
do.call(makePluginConfig, x)
}

