% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildplugin.R
\name{npmBuild}
\alias{npmBuild}
\title{Run build process to update app.min.js and app.css}
\usage{
npmBuild(pluginDir = ".", type = "nwb")
}
\description{
Run build process to update app.min.js and app.css
}

