% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{numericInput}
\alias{checkboxInput}
\alias{dropdownInput}
\alias{listInput}
\alias{numericInput}
\alias{radioInput}
\alias{selectInput}
\alias{textInput}
\title{Question Inputs}
\usage{
numericInput(x, default)

checkboxInput(x, default = FALSE)

radioInput(x, default = FALSE)

textInput(x, default = "[\\"\\"]")

listInput(x, default = "[\\"\\"]")

dropdownInput(x, default = "")

selectInput(...)
}
\arguments{
\item{x}{x}

\item{default}{default}

\item{...}{arguments to pass to selectInput}
}
\description{
Question Inputs
}

