% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.R
\name{runTests}
\alias{runTests}
\title{Run all tests in the Supporting_Macros/tests folder}
\usage{
runTests(pluginDir = ".", build_doc = TRUE)
}
\arguments{
\item{pluginDir}{plugin directory}

\item{build_doc}{whether or not to build the README file.}
}
\description{
Run all tests in the Supporting_Macros/tests folder
}

