% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWorkflow.R
\name{runWorkflows}
\alias{runWorkflows}
\title{Run Workflows in a directory}
\usage{
runWorkflows(wdir)
}
\arguments{
\item{wdir}{directory containing the workflows to run.}
}
\description{
Run Workflows in a directory
}

