% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffoldPlugin.R
\name{scaffoldPluginFromMacro}
\alias{scaffoldPluginFromMacro}
\title{Scaffold a plugin based on an existing predictive macro.}
\usage{
scaffoldPluginFromMacro(macro, ...)
}
\arguments{
\item{macro}{name of the predictive macro}
}
\description{
Scaffold a plugin based on an existing predictive macro.
}

