% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffoldWorkflow}
\alias{scaffoldWorkflow}
\title{Scaffold a new workflow based on a template}
\usage{
scaffoldWorkflow(name = "mymacro", outDir = ".",
  template = aTemplate("predictive_template.yxmc"), edit = TRUE, ...)
}
\arguments{
\item{name}{name}

\item{outDir}{output directory}

\item{template}{template}

\item{edit}{open file for editing}

\item{...}{additional arguments passed on to extractRcode}
}
\description{
Scaffold a new workflow based on a template
}
\seealso{
Other scaffold: \code{\link{scaffoldPluginFromMacro}},
  \code{\link{scaffoldPlugin}}
}

