% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlplugins.R
\name{updateHtmlPlugin}
\alias{updateHtmlPlugin}
\title{Install HTML Plugin in Alteryx}
\usage{
updateHtmlPlugin(pluginDir = ".", alteryxDir = getOption("alteryx.path"),
  build = FALSE)
}
\arguments{
\item{pluginDir}{plugin directory}

\item{alteryxDir}{alteryx directory}

\item{build}{whether or not to build js files}
}
\description{
Install HTML Plugin in Alteryx
}

