% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{updatePlugin}
\alias{updatePlugin}
\title{Update plugin}
\usage{
updatePlugin(pluginDir = ".", ...)
}
\arguments{
\item{pluginDir}{plugin directory}

\item{...}{additional arguments to pass to createPluginFromMacro}
}
\description{
This function does three things
\enumerate{
  \item Copies R code to the macro
  \item Runs \code{\link{createPluginFromMacro}} to update config.XML and gui.html
  \item Installs plugin to the htmlplugins folder in Alteryx.
}
}

