% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlplugins.R
\name{updateSvnFolder}
\alias{updateSvnFolder}
\title{Update SVN Folder}
\usage{
updateSvnFolder(pluginDir = ".", alteryxDir = "C://Desktop",
  build = FALSE)
}
\arguments{
\item{pluginDir}{plugin directory}

\item{alteryxDir}{alteryx directory}

\item{build}{whether or not to build}
}
\description{
Update SVN Folder
}

