% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.Alteryx2}
\alias{write.Alteryx2}
\title{Alteryx Write Function}
\usage{
write.Alteryx2(data, nOutput = 1, bIncludeRowNames = FALSE, source = "")
}
\arguments{
\item{data}{data}

\item{nOutput}{output connection number}

\item{bIncludeRowNames}{include row names}

\item{source}{source}
}
\description{
Alteryx Write Function
}

