% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginGui.R
\name{writeGuiHtml}
\alias{writeGuiHtml}
\title{Write Gui HTML for plugin}
\usage{
writeGuiHtml(pluginDir, htmlFile = NULL, overrides = NULL)
}
\description{
@param pluginDir directory containing the plugin.
 @param htmlFile html file to write the Gui to
 @param overrides a list of overrides for widgets
}

