% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yxmc2yaml.R
\name{yxmc2yaml}
\alias{yxmc2yaml}
\title{Extract question constants and their configuration from macro}
\usage{
yxmc2yaml(yxmc, asYaml = F)
}
\arguments{
\item{yxmc}{path to yxmc file to extract configuration from.}

\item{asYAML}{boolean indicating whether output should be converted to yaml.}
}
\description{
Extract question constants and their configuration from macro
}
\examples{
yxmc <- system.file("templates", "sample1.yxmc", package = 'AlteryxRhelper')
d <- yxmc2yaml(yxmc)
}

