% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildplugin.R
\name{buildPlugin}
\alias{buildPlugin}
\title{Build Plugin}
\usage{
buildPlugin(pluginDir = ".", build = FALSE)
}
\arguments{
\item{pluginDir}{directory containing the plugin}
}
\description{
Build Plugin
}

