% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHtmlPluginNew.R
\name{buildPlugin2}
\alias{buildPlugin2}
\title{Build Plugin}
\usage{
buildPlugin2(pluginDir = ".", build = FALSE)
}
\arguments{
\item{pluginDir}{directory containing the plugin}
}
\description{
Build Plugin
}

