% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copyPredictiveMacro}
\alias{copyPredictiveMacro}
\title{Copy predictive macro}
\usage{
copyPredictiveMacro(macro, to = ".")
}
\arguments{
\item{macro}{macro}

\item{to}{to}
}
\description{
Copy predictive macro
}

