% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayx.R
\name{createPluginFromMacro}
\alias{createPluginFromMacro}
\title{Create Plugin from Macro}
\usage{
createPluginFromMacro(pluginDir = ".", overrides = NULL, layout = NULL)
}
\arguments{
\item{pluginDir}{plugin directory}

\item{overrides}{should an override file be used}

\item{layout}{should a layout file be used}
}
\description{
Create Plugin from Macro
}

