% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createYXI}
\alias{createYXI}
\title{Create a YXI file}
\usage{
createYXI(pluginDir = ".", toDir = ".")
}
\arguments{
\item{pluginDir}{plugin directory}

\item{toDir}{to directory}
}
\description{
YXI files are zip files with a `.yxi` extension that allow a plugin to be
packaged along with all its dependencies. This function allows a user to
create a YXI file programatically. The resulting `.yxi` file can be
distributed standalone.
}

