% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlplugins.R
\name{createYXI}
\alias{createYXI}
\title{Create a YXI installable}
\usage{
createYXI(pluginDir = ".", toDir = ".")
}
\arguments{
\item{pluginDir}{directory containing the plugin files}

\item{toDir}{directory to write the yxi file to}
}
\description{
Create a YXI installable
}

