% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractInputOutput}
\alias{extractInputOutput}
\title{Extract annotation from Inputs}
\usage{
extractInputOutput(pluginName, type = "input")
}
\arguments{
\item{pluginName}{path to yxmc file}

\item{type}{whether to extract from input or output}
}
\description{
Extract annotation from Inputs
}

