% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inAlteryx}
\alias{inAlteryx}
\title{Check if the execution environment is Alteryx}
\usage{
inAlteryx()
}
\description{
It utilizes the presence of certain global variables to detect if the 
execution environment is Alteryx or not.It returns \code{TRUE} if called
inside an Alteryx R Tool.
}

