% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automation.R
\name{installAllPackages}
\alias{installAllPackages}
\title{Install all needed packages that are missing}
\usage{
installAllPackages(dev = TRUE)
}
\description{
Install all needed packages that are missing
}

