% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svnfunctions.R
\name{installToSvn}
\alias{installToSvn}
\title{Install packages to R library in SVN}
\usage{
installToSvn(srcPkg, install = TRUE)
}
\arguments{
\item{srcPkg}{path to source package}

\item{svnLibDir}{svn library directory}
}
\description{
Install packages to R library in SVN
}

