% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginconfig.R
\name{makeIcon}
\alias{makeIcon}
\title{Make a circular or square icon and save it as a png file}
\usage{
makeIcon(iconPath, shape = "circle", fill = sample(colors(), 1))
}
\arguments{
\item{iconPath}{path to save icon to}

\item{shape}{shape of the icon (circle or rect)}

\item{fill}{fill color}
}
\description{
Make a circular or square icon and save it as a png file
}

