% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWorkflow.R
\name{runWorkflow}
\alias{runWorkflow}
\title{Run an Alteryx workflow}
\usage{
runWorkflow(file)
}
\arguments{
\item{file}{parameter to pass on to AlteryxEngineCmd}
}
\description{
Run an Alteryx workflow
}

