% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffoldPlugin}
\alias{scaffoldPlugin}
\title{Scaffold a new plugin with a predictive macro}
\usage{
scaffoldPlugin(pluginName, ...)
}
\arguments{
\item{pluginName}{name of the plugin}

\item{...}{extra arguments to pass on to createPluginFromMacro}
}
\description{
Scaffold a new plugin with a predictive macro
}
\seealso{
Other scaffold: \code{\link{scaffoldPluginFromMacro}},
  \code{\link{scaffoldWorkflow}}
}

