% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffoldPluginFromMacro}
\alias{scaffoldPluginFromMacro}
\title{Scaffold a plugin based on an existing predictive macro.}
\usage{
scaffoldPluginFromMacro(macro, ...)
}
\arguments{
\item{macro}{name of the predictive macro}

\item{...}{additional arguments to pass on to createPluginFromMacro}
}
\description{
Scaffold a plugin based on an existing predictive macro.
}
\seealso{
Other scaffold: \code{\link{scaffoldPlugin}},
  \code{\link{scaffoldWorkflow}}
}

