% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaffoldWorkflow}
\alias{scaffoldWorkflow}
\title{Scaffold a new workflow based on a template}
\usage{
scaffoldWorkflow(name = "mymacro", outDir = ".",
  template = aTemplate("empty.yxmc"), edit = TRUE, ...)
}
\description{
Scaffold a new workflow based on a template
}

