% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluginconfig.R
\name{yxmc2PluginConfig}
\alias{yxmc2PluginConfig}
\title{Make plugin configuration xml file from macro}
\usage{
yxmc2PluginConfig(yxmcFile, saveToFile = NULL)
}
\arguments{
\item{yxmcFile}{path to macro}

\item{saveToFile}{file to save config.xml to}
}
\description{
Make plugin configuration xml file from macro
}

